Check that incoming packet is a SYN packet before creating a new connection.
There are some error cases where it's useful to prevent a new connection from
being created if an incoming packet is not a SYN packet. This shouldn't normally
happen, but if it does it could cause both nodes to get stuck in a loop connecting
and disconnecting to each other.