Add ability to change relay address on a connection
This commit adds the ability to change a connection's relay address without
having to close it and create a new one.

It adds a new method on the Connection class to change the connection. It will
automatically change the relay address if it receives a packet for a connection
from a different address.